Rename Command, change swerve structure
SwerveDriveCommand is now SwerveDriveJoyStick. New classes allow for the easier management of swerve drive wheels and modules. Comments were added and JavaDoc comments were updated to allow for easier use of the swerve library. Speaking of library, added a new folder for swerve library code, that can be transfered from year to year, robot to robot and not have to be created.